Avoid compiler warning

+++ app/src/main/cpp/baresip.c
@@ -1640,7 +1640,7 @@ JNIEXPORT void JNICALL Java_com_tutpro_baresip_plus_Api_call_1set_1media_1direct
 {
     (void)env;
     (void)obj;
-    LOGD("call set audio/video media direction of call %ld to %d/%d\n", call, adir, vdir);
+    LOGD("call set audio/video media direction of call %ld to %d/%d\n", (long)call, adir, vdir);
     re_thread_enter();
     call_set_media_direction((struct call *)call, (enum sdp_dir)adir, (enum sdp_dir)vdir);
     re_thread_leave();
